requirements.txt
# Dépendances existantes
requests==2.31.0
python-dotenv==1.0.0
groq>=0.4.2

# Nouvelles dépendances pour RAG
chromadb==0.4.22
sentence-transformers==2.5.1
numpy==1.26.4
pandas==2.2.1
torch==2.2.1
pydantic==2.6.3

# Utilitaires
tqdm==4.66.2
pytest==8.0.2
tenacity==8.2.3
pytest-asyncio==0.23.6

# Backends LLM supplémentaires
openai>=1.18.0
anthropic>=0.18.1 